Fixed another compiler warning.

+++ source/core/board/NstBoardSunsoft5b.cpp
@@ -364,7 +364,7 @@ namespace Nes
 					status = data & (0x1|0x8);
 
 					if (status & 0x1)
-						dc = ~0UL;
+						dc = (unsigned int) ~0UL;
 				}
 
 				void S5b::Sound::Square::WriteReg3(const uint data)
